requirements.txt
Django==1.11.3
django-jquery==3.1.0
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
pytz==2017.2
